[qr_scan] rename qrScanPage
