+++ lib/mocks/qrScanPage.dart
