+++ lib/page/account/uos/qrSenderPage.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 
 import 'package:encointer_wallet/common/components/roundedButton.dart';
-import 'package:encointer_wallet/page/qr_scan/scanPage.dart';
+import 'package:encointer_wallet/page/qr_scan/qrScanPage.dart';
 import 'package:encointer_wallet/service/js_api_wrappers/api.dart';
 import 'package:encointer_wallet/utils/translations/index.dart';
 import 'package:flutter/cupertino.dart';
