+++ lib/page/qr_scan/qrScanPage.dart
