+++ test_driver/scanPage.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:encointer_wallet/mocks/restartWidget.dart';
-import 'package:encointer_wallet/mocks/scanPage.dart';
+import 'package:encointer_wallet/mocks/qrScanPage.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_driver/driver_extension.dart';
 import 'package:rxdart/rxdart.dart';
